bug fix: update tables correctly when a router dies
The costs of routers were not being updated correctly when the one neighbor dies. The reason was that the NodeRouter used to incorrectly update the static neighbor link cost, because other neighbor broadcasts smaller cost.
e.g. With router B, the "static" cost of neighbor C is 3. And through neighbor F, it’s 2. So on broadcast by F, B updates the static cost in its table for router C as 2. Since it doesn’t load the static cost from file again, when F dies, the cost stays 2 to neighbor C directly.

Now, it uses ref_cost even for neighbor nodes, and router->cost is used to store original cost which is static.